+++ libcnb-package/src/buildpack_dependency.rs
@@ -1,7 +1,6 @@
+use crate::output::BuildpackOutputDirectoryLocator;
 use libcnb_data::buildpack::{BuildpackId, BuildpackIdError};
 use libcnb_data::buildpackage::{Buildpackage, BuildpackageDependency};
-use std::collections::HashMap;
-use std::hash::BuildHasher;
 use std::path::{Path, PathBuf};
 
 /// Buildpack dependency type
@@ -80,9 +79,9 @@ pub fn get_local_buildpackage_dependencies(
 /// * the given `buildpackage` contains a local dependency with an invalid [`BuildpackId`]
 /// * there is no entry found in `buildpack_ids_to_target_dir` for a local dependency's [`BuildpackId`]
 /// * the target path for a local dependency is an invalid URI
-pub fn rewrite_buildpackage_local_dependencies<S: BuildHasher>(
+pub fn rewrite_buildpackage_local_dependencies(
     buildpackage: &Buildpackage,
-    buildpack_ids_to_target_dir: &HashMap<&BuildpackId, PathBuf, S>,
+    buildpack_output_directory_locator: &BuildpackOutputDirectoryLocator,
 ) -> Result<Buildpackage, RewriteBuildpackageLocalDependenciesError> {
     let local_dependency_to_target_dir = |target_dir: &PathBuf| {
         BuildpackageDependency::try_from(target_dir.clone()).map_err(|_| {
@@ -99,14 +98,10 @@ pub fn rewrite_buildpackage_local_dependencies<S: BuildHasher>(
                     BuildpackDependency::External(buildpackage_dependency) => {
                         Ok(buildpackage_dependency)
                     }
-                    BuildpackDependency::Local(buildpack_id, _) => buildpack_ids_to_target_dir
-                        .get(&buildpack_id)
-                        .ok_or(
-                            RewriteBuildpackageLocalDependenciesError::TargetDirectoryLookup(
-                                buildpack_id,
-                            ),
-                        )
-                        .and_then(local_dependency_to_target_dir),
+                    BuildpackDependency::Local(buildpack_id, _) => {
+                        let output_dir = buildpack_output_directory_locator.get(&buildpack_id);
+                        local_dependency_to_target_dir(&output_dir)
+                    }
                 })
                 .collect()
         })
@@ -120,7 +115,6 @@ pub fn rewrite_buildpackage_local_dependencies<S: BuildHasher>(
 /// An error for [`rewrite_buildpackage_local_dependencies`]
 #[derive(Debug)]
 pub enum RewriteBuildpackageLocalDependenciesError {
-    TargetDirectoryLookup(BuildpackId),
     InvalidDependency(PathBuf),
     GetBuildpackDependenciesError(BuildpackIdError),
 }